+++ .gitignore
@@ -1,3 +1,1 @@
 .DS_Store
-node_modules/
-dist/
