--- webpack.config.js
@@ -1,62 +1,0 @@
-const path = require("path");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const { CleanWebpackPlugin } = require("clean-webpack-plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
-
-module.exports = {
-    entry: {
-        main: "./src/pages/index.js",
-    },
-    output: {
-        path: path.resolve(__dirname, "dist"),
-        filename: "main.js",
-        publicPath: "",
-    },
-
-    mode: "development",
-    devtool: "inline-source-map",
-    stats: "errors-only",
-    devServer: {
-        static: path.resolve(__dirname, "./dist"),
-        compress: true,
-        port: 8080,
-        open: true,
-        liveReload: true,
-        hot: false,
-    },
-    target: ["web", "es5"],
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                loader: "babel-loader",
-                exclude: "/node_modules/",
-            },
-            {
-                test: /\.css$/,
-                use: [
-                    MiniCssExtractPlugin.loader,
-                    {
-                        loader: "css-loader",
-                        options: {
-                            importLoaders: 1,
-                        },
-                    },
-                    "postcss-loader",
-                ],
-            },
-            {
-                test: /\.(png|svg|jpg|jpeg|webp|gif|woff(2)?|eot|ttf|otf)$/,
-                type: "asset/resource",
-            },
-        ],
-    },
-    plugins: [
-        new HtmlWebpackPlugin({
-            template: "./src/index.html",
-            favicon: "./src/images/favicon.ico",
-        }),
-        new CleanWebpackPlugin(),
-        new MiniCssExtractPlugin(),
-    ],
-};
